+++ arm9/source/romBrowser/IconCoverMode.h
@@ -1,0 +1,9 @@
+#pragma once
+
+/// @brief The way icon fallback covers are shown in coverflow.
+enum class IconCoverMode
+{
+    Disabled,
+    Original,
+    Large
+};