+++ arm9/source/themes/ThemeInfo.h
@@ -3,6 +3,7 @@
 #include "ThemeType.h"
 #include "core/String.h"
 #include "core/math/Rgb.h"
+#include "romBrowser/IconCoverMode.h"
 
 class ThemeInfo
 {
@@ -13,20 +14,24 @@ class ThemeInfo
     String<char16_t, 64> _author;
     Rgb<8, 8, 8> _primaryColor;
     bool _darkTheme;
+    IconCoverMode _iconCoverMode;
+    Rgb<8, 8, 8> _iconCoverBackgroundColor;
 public:
     ThemeInfo(const TCHAR* folderName, ThemeType type, const char* name,
         const char* description, const char* author, const Rgb<8, 8, 8>& primaryColor,
-        bool darkTheme)
+        bool darkTheme, IconCoverMode iconCoverMode, const Rgb<8, 8, 8>& iconCoverBackgroundColor)
         : _folderName(folderName), _type(type), _name(name)
         , _description(description), _author(author), _primaryColor(primaryColor)
-        , _darkTheme(darkTheme) { }
+        , _darkTheme(darkTheme), _iconCoverMode(iconCoverMode)
+        , _iconCoverBackgroundColor(iconCoverBackgroundColor) { }
 
     ThemeInfo(const TCHAR* folderName, ThemeType type, const char16_t* name,
         const char16_t* description, const char16_t* author, const Rgb<8, 8, 8>& primaryColor,
-        bool darkTheme)
+        bool darkTheme, IconCoverMode iconCoverMode, const Rgb<8, 8, 8>& iconCoverBackgroundColor)
         : _folderName(folderName), _type(type), _name(name)
         , _description(description), _author(author), _primaryColor(primaryColor)
-        , _darkTheme(darkTheme) { }
+        , _darkTheme(darkTheme), _iconCoverMode(iconCoverMode)
+        , _iconCoverBackgroundColor(iconCoverBackgroundColor) { }
 
     constexpr const TCHAR* GetFolderName() const { return _folderName; }
     constexpr ThemeType GetType() const { return _type; }
@@ -35,4 +40,6 @@ public:
     constexpr const char16_t* GetAuthor() const { return _author; }
     constexpr const Rgb<8, 8, 8>& GetPrimaryColor() const { return _primaryColor; }
     constexpr bool GetIsDarkTheme() const { return _darkTheme; }
+    constexpr IconCoverMode GetIconCoverMode() const { return _iconCoverMode; }
+    constexpr const Rgb<8, 8, 8>& GetIconCoverBackgroundColor() const { return _iconCoverBackgroundColor; }
 };